new structure for email
