+++ server/components/email/index.js
@@ -1,48 +1,28 @@
+/**
+ * Email Module. The email service to use is determined by the emailService
+ * in the config.
+ */
+
+/**
+ * Sends the initial signup email.
+ * @param  {object}   message  {email, name, verifyURL}
+ * @param  {Function} callback Function to call after queuing or sending email
+ */
+module.exports.signup = (message, callback) => console.error("No Email Service!");
+
+/**
+ * Sends a "forgot password" email.
+ * @param  {object}   message  {email, name, verifyURL}
+ * @param  {Function} callback Function to call after queuing or sending email
+ */
+module.exports.forgotPassword = (message, callback) => console.error("No Email Service!");
+
 var config = require('../../config/environment');
-var sendgrid = require("sendgrid")(config.sendgridKey);
-var fs = require('fs');
-var path = require('path');
-var signupTemplate = fs.readFileSync(path.join(__dirname,"signup.html"), "utf8");
-var forgotPasswordTemplate = fs.readFileSync(path.join(__dirname,"forgotPassword.html"), "utf8");
 
-module.exports.signup = function(message, callback){
-  callback = callback || function(){};
-  var email = new sendgrid.Email();
-  email.addTo(message.email);
-  email.subject = "[Venue] Signup Verification";
-  email.from = config.serverEmail;
-  email.setHtml(signupTemplate); //pass in the string template we read from disk
-  email.addSubstitution("-name-", message.name); //sub. variables
-  email.addSubstitution("-verifyURL-", message.verifyURL); //sub. variables
-  sendEmail(email, callback)
-  if(process.env.NODE_ENV != "production") {
-    console.log(message.verifyURL);
-  }
+if (config.emailService == "SMTP"){
+  module.exports = require("./smtp");
+}else if (config.emailService == "SENDGRID"){
+  module.exports = require("./sendgrid");
+}else if (config.emailService == "MOCK"){
+  module.exports = require("./mock");
 }
-
-module.exports.forgotPassword = function(message, callback){
-  callback = callback || function(){};
-  var email = new sendgrid.Email();
-  email.addTo(message.email);
-  email.subject = "[Venue] Forgot Email";
-  email.from = config.serverEmail;
-  email.setHtml(forgotPasswordTemplate); //pass in the string template we read from disk
-  email.addSubstitution("-name-", message.name); //sub. variables
-  email.addSubstitution("-verifyURL-", message.verifyURL); //sub. variables
-  sendEmail(email, callback)
-  if(process.env.NODE_ENV != "production") {
-    console.log(message.verifyURL);
-  }
-}
-
-
-function sendEmail(email, callback){
-    sendgrid.send(email, function(err, json){
-        if (err){
-          callback("An error occurred sending the email");
-          console.error(err, json);
-        }else{
-          callback(null);
-        }
-    });
-};
