+++ server/components/email/templates/forgotPassword.html
@@ -7,7 +7,7 @@
     <div>&nbsp;</div>
     <div>
       You forgot your password.
-      <a href=-verifyURL->Click here to reset</a>
+      <a href="-verifyURL-">Click here to reset</a>
     </div>
     <div>&nbsp;</div>
     <div>Good Luck</div>
