+++ server/components/email/templates/signup.html
@@ -9,6 +9,6 @@
     <div>&nbsp;</div>
     <div>Best - Team Venue</div>
     <div>&nbsp;</div>
-    <a href=-verifyURL->Click here to verify</a>
+    <a href="-verifyURL-">Click here to verify</a>
   </body>
 </html>
