+++ server/config/environment/development.js
@@ -12,6 +12,9 @@ module.exports = {
   // Images Folder
   imageUploadPath: './data/',
 
+  // Just log outgoing emails
+  emailService: "MOCK",
+
   // Seed database on startup
   seedDB: true
